minor bugs fixed and functionality of sharing data among components shifted to home.component.ts
