+++ frontend/src/app/home/home.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormControl, FormBuilder,  Validators } from '@angular/forms';
-import { Router } from '@angular/router';
 import { TransactionDetailService } from '../service/transaction-detail.service';
+import { Router, NavigationExtras } from '@angular/router';
+
 @Component({
   selector: 'app-home',
   templateUrl: './home.component.html',
@@ -16,6 +17,17 @@ export class HomeComponent implements OnInit {
   }
   async getTransactionDetails(){
     var transactionId = this.transactionForm.value.transactionId
-     var data = await this.transactionDetailService.get(transactionId)
+    await this.transactionDetailService.get(transactionId)
+    .subscribe((res: any) => {
+      let navigationExtras: NavigationExtras = {
+        queryParams: {
+            "transactionData": JSON.stringify(res)
+        }
+      };
+      this.router.navigate(["/transaction"],  navigationExtras);
+    },
+    error  => {
+      console.log(error)
+    })
   }
 }
