--- frontend/src/app/transaction/transaction.component.css
@@ -1,3 +1,0 @@
-table{
-    width: 100%;
-}
