+++ frontend/src/app/transaction/transaction.component.html
@@ -1,7 +1,7 @@
     <p></p>
     <table mat-table [dataSource]="transactionData" class="mat-elevation-z8 z-depth-5">
         <ng-container matColumnDef="class">
-            <th mat-header-cell *matHeaderCellDef class="text-right">Class </th>
+            <th mat-header-cell *matHeaderCellDef class="text-right"> Class </th>
             <td mat-cell *matCellDef="let transactionData" class="text-right"> {{transactionData.class}} </td>
         </ng-container>
         <ng-container matColumnDef="sku">
